requirements.txt
langchain>=0.3.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.75.0
python-dotenv>=1.0.0
faiss-cpu>=1.8.0
flask>=3.0.0 